refactor : miscellaneous fix
